update-patches.sh: fix issue when no existing patches again

+++ update-patches.sh
@@ -39,7 +39,9 @@ orig_patches=$(awk '/^Patch[0-9][0-9]*:/ { print $2 }' "${spec}")
 #
 # Create a commit which removes all the patches
 #
-git rm ${orig_patches}
+if [ "${orig_patches}" ]; then
+    git rm ${orig_patches}
+fi
 git commit --allow-empty -m "Updated patches from ${patches_branch}" ${orig_patches}
 
 #
